smooth triangles, broken groups

+++ apps/raytrace1.cpp
@@ -16,15 +16,15 @@ int main() {
   const auto origin = Tuple::point(0, 0, -5);
   int wall_z = 10;
 
-  float wall_size = 7.0;
-  float pixel_size = wall_size / canvas_pixels;
-  float half = wall_size / 2;
+  double wall_size = 7.0;
+  double pixel_size = wall_size / canvas_pixels;
+  double half = wall_size / 2;
 
   for (int y = 0; y < canvas_pixels - 1; ++y) {
-    float world_y = half - pixel_size * y;
+    double world_y = half - pixel_size * y;
 
     for (int x = 0; x < canvas_pixels - 1; ++x) {
-      float world_x = -half + pixel_size * x;
+      double world_x = -half + pixel_size * x;
 
       const auto position = Tuple::point(world_x, world_y, wall_z);
       const auto direction = position - origin;
