+++ tests/test_common.cpp
@@ -3,7 +3,7 @@
 #include "../core/matrix.h"
 #include "../core/tuple.h"
 
-bool vector_is_near(Tuple a, Tuple b, float abs) {
+bool vector_is_near(Tuple a, Tuple b, double abs) {
   std::cout << "A: " << a << std::endl;
   std::cout << "B: " << b << std::endl;
 
@@ -16,7 +16,7 @@ bool tuple_is_near(Tuple a, Tuple b) {
          abs(a.x - b.x) < EPSILON && abs(a.w - b.w) < EPSILON;
 }
 
-bool matrix_is_near(Matrix a, Matrix b, float abs) {
+bool matrix_is_near(Matrix a, Matrix b, double abs) {
   for (size_t row = 0; row < a.size(); ++row) {
     for (size_t col = 0; col < a.size(); ++col) {
       if (std::abs(a.get(row, col) - b.get(row, col)) > abs) {
@@ -27,7 +27,7 @@ bool matrix_is_near(Matrix a, Matrix b, float abs) {
   return true;
 }
 
-bool point_is_near(Tuple a, Tuple b, float abs) {
+bool point_is_near(Tuple a, Tuple b, double abs) {
   std::cout << "A: " << a << std::endl;
   std::cout << "B: " << b << std::endl;
