+++ tests/test_common.h
@@ -9,7 +9,7 @@
 #include "../core/matrix.h"
 #include "../core/tuple.h"
 
-bool vector_is_near(Tuple a, Tuple b, float abs);
+bool vector_is_near(Tuple a, Tuple b, double abs);
 bool tuple_is_near(Tuple a, Tuple b);
-bool matrix_is_near(Matrix a, Matrix b, float abs);
-bool point_is_near(Tuple a, Tuple b, float abs);
+bool matrix_is_near(Matrix a, Matrix b, double abs);
+bool point_is_near(Tuple a, Tuple b, double abs);
